bench: v02-only bring-up loop; fix v1 gating dtype; use --xla_dump_to
Strip v1 from the VM runner: bring-up needs v0.2 compile/dump/timing
signal and v1's lowering failure was killing the process before v0.2
executed. v1 returns for the final comparison. Its crash is fixed for
then: the bench fed f32 gating into v1's act-dtype staging buffer and
DMAs do not convert (tpu.enqueue_dma element type mismatch) - cast to
the activation dtype; the f32 copy stays for the reference and the
tuner's max-load math. The mosaic dump flag becomes the generic
--xla_dump_to (installed libtpu rejects --xla_mosaic_dump_to).

+++ tpu_inference/kernels/fused_moe/v2/bench_decode.py
@@ -171,7 +171,10 @@ def main() -> None:
             NamedSharding(mesh_v1, P("model", None, None, None)))
         w2_v1 = jax.device_put(w2, NamedSharding(mesh_v1, P("model")))
         tok_v1 = jax.device_put(tokens, NamedSharding(mesh_v1, P("model")))
-        gate_v1 = jax.device_put(gating, NamedSharding(mesh_v1, P("model")))
+        # v1 stages gating into an act-dtype VMEM buffer; f32 gating fails
+        # its DMA lowering (src/dst element type mismatch)
+        gate_v1 = jax.device_put(gating.astype(dtype),
+                                 NamedSharding(mesh_v1, P("model")))
 
         def run_v1() -> jax.Array:
             return fused_ep_moe(
